ROSE/requirements.txt
Pillow
playsound
pypiwin32